DOCS: Rename baseline folder to baselineprocess This was needed to avoid problem with the case insensitivity on MacOS. difference between baseline and Baseline caused problems during the generation of docs
