+++ spectrochempy/analysis/baselineprocessing/api.py
