+++ spectrochempy/analysis/baselineprocessing/baselineutils.py
