+++ tests/test_analysis/test_baseline/test_baseline.py
@@ -10,7 +10,7 @@ import os
 import pytest
 
 import spectrochempy as scp
-from spectrochempy.analysis.baseline.baseline import Baseline
+from spectrochempy.analysis.baselineprocessing.baseline import Baseline
 from spectrochempy.core.units import ur
 from spectrochempy.utils.plots import show
 from spectrochempy.utils.testing import assert_dataset_equal
